+++ lib/ddtrace/ext/metrics.rb
@@ -1,0 +1,11 @@
+module Datadog
+  module Ext
+    module Metrics
+      TAG_LANG = 'language'.freeze
+      TAG_LANG_INTERPRETER = 'language-interpreter'.freeze
+      TAG_LANG_VERSION = 'language-version'.freeze
+      TAG_RUNTIME_ID = 'runtime-id'.freeze
+      TAG_TRACER_VERSION = 'tracer-version'.freeze
+    end
+  end
+end